Refactor to make things easier to import
