+++ pkg/datarepair/checker/checker.go
@@ -1,4 +1,4 @@
 // Copyright (C) 2018 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package repairer
+package checker
