+++ pkg/datarepair/repair.go
@@ -4,14 +4,39 @@
 package datarepair
 
 import (
-	"storj.io/storj/pkg/pb"
+	"context"
+	"gopkg.in/spacemonkeygo/monkit.v2"
+
+	q "storj.io/storj/pkg/datarepair/queue"
+	// "storj.io/storj/pkg/datarepair/checker"
+	"storj.io/storj/pkg/datarepair/repairer"
 )
 
-//RepairQueue is the interface for the data repair queue
-type RepairQueue interface {
-	Add(qi *pb.InjuredSegment) error
-	AddAll(qis []*pb.InjuredSegment) error
-	Remove(qi *pb.InjuredSegment) error
-	GetNext() pb.InjuredSegment
-	GetSize() int
+var (
+	mon = monkit.Package()
+)
+
+// Config contains configurable values for repairer
+type Config struct {
+	//TODO: Add things for checker
+	//TODO: Add things for repairer
+}
+
+// Run runs the repairer with configured values
+func (c *Config) Run(ctx context.Context) (err error) {
+	defer mon.Task()(&ctx)(&err)
+
+	var queue q.RepairQueue
+
+	// TODO: Initialize Checker with queue
+
+	// Initialize Repairer with queue
+	_, err = repairer.Initialize(ctx, queue)
+
+	// TODO: Run the Checker in goroutine
+	// TODO: Run the Repairer in goroutine
+
+	// TODO: defer stop of checker and repairer
+
+	return nil
 }
